fix migration after model was removed some days ago

+++ src/api/db/migrate/20100629095208_drop_bs_role.rb
@@ -1,6 +1,6 @@
 class DropBsRole < ActiveRecord::Migration
   def self.up
-    drop_table "bs_roles"
+    drop_table "bs_roles" if ActiveRecord::Base.connection.tables.include?("bs_roles")
   end
 
   def self.down
